+++ client/src/components/RightSidebar/RightSidebar.module.scss
@@ -13,10 +13,11 @@
   max-height: 52vh;
   background-color: $background-secondary;
   border-radius: 10px;
-  overflow-y: auto; 
-  transition: height 1s ease;   
+  overflow-y: auto;
+  transition: height 1s ease;
   border: 3px solid $twelfth;
-  &.expanded {    
+
+  &.expanded {
     max-height: 52vh;
   }
 
@@ -38,9 +39,9 @@
   }
 
   @media (max-width: 767px) {
-    display: none;
-    border: 1px solid red
+    border: 1px solid red; /* Adjust or remove if not needed */
   }
+
   &.showInMainContent {
     display: flex !important;
   }
@@ -51,78 +52,3 @@
     }
   }
 }
-
-.friendItem {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 10px;
-  padding: 10px;
-  width: 100%;
-}
-
-.friendLink {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 11vh;
-  gap: 10px;
-  text-decoration: none;
-  color: $secondary;
-  border: solid 1px $secondary;
-  border-radius: 20px;
-  padding: 5px;
-  flex-grow: 1;
-
-  &:hover {
-    color: $background-btn-fifth;
-    transform: scale(1.1);
-    box-shadow: 0 0 15px rgba(83, 83, 83, 0.8), 0 0 25px rgba(52, 53, 53, 0.5);
-  }
-}
-.toggleButton span {
-  margin-right: 10px;
-}
-.toggleButton {
-  color: $primary;
-  font-size: $medium;
-  padding: 5px 10px;
-  border-radius: 10px;
-  background-color: $background-btn-fifth;
-}
-.toggleButton:hover {
-  background-color: $background-btn-hover-fifth;
-}
-
-.friendName {
-  display: flex;
-  flex-direction: column;
-}
-
-.friendName,
-.lastName p {
-  font-size: $large;
-}
-
-.deleteButton {
-  margin-left: auto;
-}
-
-.friendAvatar {
-  width: 70px;
-  height: 70px;
-  border-radius: 50%;
-}
-
-.friendsTitle {
-  position: sticky;
-  top: 0; 
-  z-index: 10; 
-  padding: 5px 0;
-  margin-top: 0;
-  font-size: 1.5rem; 
-  width: 100%;
-  background-color: $background-secondary;
-  text-align: center;
-}
-
